requirements.txt
anthropic==0.36.2
beautifulsoup4==4.12.3
boto3==1.35.44
colormath==3.0.0
datasets==2.19.0
joblib==1.3.2
matplotlib==3.8.1
numpy==1.22.4
openai==1.52.0
openai_clip==1.0.1
opencv_contrib_python==4.6.0.66
opencv_python==4.6.0.66
opencv_python_headless==4.9.0.80
paddleocr==2.7.3
Pillow==10.4.0
playwright==1.43.0
protobuf==5.28.2
retry==0.9.2
scikit_learn==1.4.0
scipy==1.13.1
Shapely==2.0.6
streamlit==1.39.0
torch==2.4.0
torchvision==0.19.0
tqdm==4.66.2
transformers==4.44.1
accelerate=1.0.1
bitsandbytes=0.44.1